takos hostの共通化
