+++ app/takos_host/README.md
@@ -12,14 +12,14 @@ takos を運用できるようにすることが目的です。
    サーバーを作成します。
 2. サーバーごとに専用のドメインを割り当て、Deno
    プロセスとして起動します。仮想化は行わず、すべての処理を Deno
-   プロセスで完結させます。
-   takosを利用してたくさん起動して成立するようなソフトウェアではなく、
-   takosと同じ動きのするインスタンスを単一のスクリプトでさばけるようにするソフトウェアです。
+   プロセスで完結させます。takos
+   を利用してたくさん起動して成立するようなソフトウェアではなく、takos
+   と同じ動きのするインスタンスを単一のスクリプトでさばけるようにするソフトウェアです。
 3. 各インスタンスはデフォルトで relay.takos.jp
    のリレーに参加しており、内部的にはリレーに近い仕組みを inbox
    を介さずに実装します。
 4. ユーザーは発行されたドメインへアクセスし、パスワードを入力してログインします。
 
-
 ホスト側ではドメイン名を基にインスタンスを識別し、takos の API
-を抽象化して処理します。これにより複数の takos サーバーを一元管理できます。
+を共通モジュールとして読み込んで処理します。これにより複数の takos
+サーバーを一元管理できます。
