+++ app/takos_host/deno.json
@@ -4,6 +4,15 @@
   },
   "imports": {
     "@std/assert": "jsr:@std/assert@1",
-    "hono": "npm:hono@^4.8.4"
+    "hono": "npm:hono@^4.8.4",
+    "@aws-sdk/util-stream-node": "npm:@aws-sdk/util-stream-node@^3.374.0",
+    "@hono/zod-validator": "npm:@hono/zod-validator@^0.7.0",
+    "@std/dotenv": "jsr:@std/dotenv@^0.225.5",
+    "@std/fs": "jsr:@std/fs@^1.0.18",
+    "@std/path": "jsr:@std/path@^1.1.0",
+    "mongoose": "npm:mongoose@^8.15.1",
+    "zod": "npm:zod@^3.25.56",
+    "@aws-sdk/client-s3": "npm:@aws-sdk/client-s3@^3.528.0",
+    "mongodb": "npm:mongodb@^6.17.0"
   }
 }
